container/requirements.txt
opencv-python
numpy
flask
pillow
gunicorn
requests==2.22.0
fastai
